Added all resful api routes to apiroutes blueprint
