+++ api/blueprints/apiroutes.py
@@ -1,0 +1,77 @@
+import json
+import re
+from api.config import login_manager, year, domain
+from api.blueprints.information import info_bp
+from api.blueprints.authentication import auth_bp
+from datetime import datetime, timezone, timedelta
+from flask import Blueprint, flash, request, session, render_template, url_for, redirect, jsonify
+from flask_login import current_user, login_required, login_user, logout_user
+from math import ceil
+from models.result import Result
+from models.user import User
+from models.quiz import Quiz
+from models import quizzesCollection, resultsCollection
+from urllib.parse import urlparse
+from pprint import pprint
+
+api_db = Blueprint('api_db', __name__)
+
+
+@api_db.route('/get', methods=['GET'])
+def getAll():
+    """ Gets all quizzes from database
+    """
+    page = int(request.args.get('page', 1))
+    if page <= 0:
+        page = 1
+
+    per_page = int(request.args.get('per_page', 30))
+    if per_page > 100:
+        per_page = 100
+    if per_page <= 0:
+        per_page = 30
+
+    skip = (page - 1) * per_page
+    query = request.args.get('search', '')
+
+    pattern = re.compile(r'[^a-zA-Z0-9\s]')
+    if pattern.search(query):
+        query = re.escape(query)
+
+    if query:
+        total = quizzesCollection.count_documents(
+            {"title": {"$regex": query, "$options": "i"}}
+            )
+        total_pages = ceil(total / per_page) if total > 0 else 1
+        cursor = quizzesCollection.find(
+            {"title": {"$regex": query, "$options": "i"}},
+            {"_id": False, "creator_id": False},
+            ).sort([("title", 1), ("updated_at", 1)]).skip(skip).limit(per_page)
+        quizzes = list(cursor)
+    else:
+        total = quizzesCollection.count_documents({})
+        total_pages = ceil(total / per_page) if total > 0 else 1
+        # cursor = Quiz.getAll().skip(skip).sort("updated_at", 1).limit(per_page)
+        cursor = quizzesCollection.find({}, {'_id': False, 'creator_id': False}).skip(skip).sort("updated_at", 1).limit(per_page)
+        quizzes = list(cursor)
+
+    if quizzes:
+        return jsonify(quizzes)
+    else:
+        return jsonify({
+            "message": "Not found"
+        }), 500
+
+
+@api_db.route('/get/<quiz_id>', methods=["GET"])
+def get(quiz_id):
+    """ Gets a quiz and returns it as json
+    """
+    cursor = quizzesCollection.find({"quiz_id": quiz_id}, {"_id": False, "creator_id": False})
+    quiz = list(cursor)
+    if not quiz:
+        return jsonify({
+            "message": "Quiz not found"
+        }), 400
+
+    return jsonify(quiz)
